backend/requirements.txt
Django==4.1.2
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djoser==2.1.0
phonenumbers==8.12.57
Pillow==9.2.0
psycopg2==2.9.4
python-dotenv==0.21.0
drf-spectacular==0.24.2
